requirements.txt
fastapi
uvicorn[standard]
asyncpg
aiohttp
playwright
pymupdf
python-docx
python-pptx
openpyxl
